% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{adj.rle}
\alias{adj.rle}
\title{Adjacent Values Run Length Encoding}
\usage{
adj.rle(x)
}
\arguments{
\item{x}{a numeric vector with}
}
\value{
a list with two elements: vector of run lengths, and another list of values corresponding to generated sequences' lengths.
}
\description{
Similar to \code{\link{rle}} function, this function detects "runs" of adjacent integers, and displays vector of run lengths and list of corresponding integer sequences.
}
\references{
See original thread for more details \url{https://stackoverflow.com/a/8467446/564164}. Special thanks to Gabor Grothendieck for this one!
}
\author{
Gabor Grothendieck <ggrothendieck@gmail.com>
}
