% Generated by roxygen2 (4.0.2): do not edit by hand
\name{capitalise}
\alias{capitalise}
\title{Capitalise String}
\usage{
capitalise(x)
}
\arguments{
\item{x}{a character vector to capitalise}
}
\value{
character vector with capitalised string elements
}
\description{
Capitalises strings in provided character vector
}
\examples{
capitalise(c("foo", "bar")) # [1] "Foo" "Bar"
}

