% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textual.R
\name{fraction.to.string}
\alias{fraction.to.string}
\title{A fraction in ordinary English language}
\usage{
fraction.to.string(x)
}
\arguments{
\item{x}{numeric}
}
\value{
string
}
\description{
A fraction in ordinary English language
}
