\name{max}
\alias{max}
\alias{rp.max}
\title{Maximum}
\usage{
max(...)
}
\arguments{
  \item{...}{parameters to be passed to \code{univar}
  function}
}
\value{
a numeric value with maximum value
}
\description{
Returns the maximum of all values in a vector by passing
\{code{\link{max}} as \code{fn} argument to
\code{\link{univar}} function.
}

