% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univar.R
\name{min}
\alias{min}
\alias{rp.min}
\title{Minimum}
\usage{
min(...)
}
\arguments{
\item{...}{parameters to be passed to \code{univar} function}
}
\value{
a numeric value with minimum value
}
\description{
Returns the minimum of all values in a vector by passing \{code{\link{min}} as \code{fn} argument to \code{\link{univar}} function.
}
