\name{range}
\alias{range}
\alias{rp.range}
\title{Range}
\usage{
range(...)
}
\arguments{
  \item{...}{parameters to be passed to \code{univar}
  function}
}
\value{
a numeric value with calculated range
}
\description{
Calculates difference between the largest and the smallest
value in a vector. See \code{\link{univar}} for details.
}

