\name{se.mean}
\alias{rp.se.mean}
\alias{se.mean}
\title{Standard Error of Mean}
\usage{
se.mean(...)
}
\arguments{
  \item{...}{parameters to be passed to \code{univar}
  function}
}
\value{
a numeric value with standard error of mean
}
\description{
Calculates standard error of mean for given variable. See
\code{\link{univar}} for details.
}

