% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textual.R
\name{synonym}
\alias{synonym}
\title{Synonym}
\usage{
synonym(word)
}
\arguments{
\item{word}{a word to look-up in `rapportools::synonyms`}
}
\value{
a synonym if found in `rapportools::synonyms` words
}
\description{
Synonym
}
\examples{
synonym('package')
synonym('bar')
}
