\name{iqr}
\alias{IQR}
\alias{iqr}
\alias{rp.iqr}
\title{Interquartile Range}
\usage{
iqr(...)
}
\arguments{
  \item{...}{parameters to be passed to \code{univar}
  function}
}
\value{
a numeric value with variable's interquartile range
}
\description{
Calculates interquartile range of given variable. See
\code{\link{univar}} for details.
}

