% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univar.R
\name{iqr}
\alias{iqr}
\alias{IQR}
\alias{rp.iqr}
\title{Interquartile Range}
\usage{
iqr(...)
}
\arguments{
\item{...}{parameters to be passed to \code{univar} function}
}
\value{
a numeric value with variable's interquartile range
}
\description{
Calculates interquartile range of given variable. See \code{\link{univar}} for details.
}
