% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univar.R
\name{median}
\alias{median}
\alias{rp.median}
\title{Median}
\usage{
median(...)
}
\arguments{
\item{...}{parameters to be passed to \code{univar} function}
}
\value{
a numeric value with variable's median
}
\description{
Calculates median of given variable. See \code{\link{univar}} for details.
}
