\name{sd}
\alias{rp.sd}
\alias{sd}
\title{Standard Deviation}
\usage{
sd(...)
}
\arguments{
  \item{...}{parameters to be passed to \code{univar}
  function}
}
\value{
a numeric value with variable's standard deviation
}
\description{
Calculates standard deviation of given variable. See
\code{\link{univar}} for details.
}

