% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univar.R
\name{sd}
\alias{sd}
\alias{rp.sd}
\title{Standard Deviation}
\usage{
sd(...)
}
\arguments{
\item{...}{parameters to be passed to \code{univar} function}
}
\value{
a numeric value with variable's standard deviation
}
\description{
Calculates standard deviation of given variable. See \code{\link{univar}} for details.
}
