% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univar.R
\name{se.mean}
\alias{se.mean}
\alias{rp.se.mean}
\title{Standard Error of Mean}
\usage{
se.mean(...)
}
\arguments{
\item{...}{parameters to be passed to \code{univar} function}
}
\value{
a numeric value with standard error of mean
}
\description{
Calculates standard error of mean for given variable. See \code{\link{univar}} for details.
}
