\name{sum}
\alias{rp.sum}
\alias{sum}
\title{Sum}
\usage{
sum(...)
}
\arguments{
  \item{...}{parameters to be passed to \code{univar}
  function}
}
\value{
a numeric value with sum of vector elements
}
\description{
Returns the sum of variable's elements, by passing
\code{\link{sum}} as \code{fn} argument to
\code{\link{univar}} function.
}

