% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univar.R
\name{var}
\alias{var}
\alias{rp.var}
\title{Variance}
\usage{
var(...)
}
\arguments{
\item{...}{parameters to be passed to \code{univar} function}
}
\value{
a numeric value with variable's variance
}
\description{
Calculates variance of given variable. See \code{\link{univar}} for details.
}
